optimalisasi coding

+++ app/src/main/java/me/dzikry/movapp/data/models/Movie.kt
@@ -8,6 +8,6 @@ data class Movie(
     @SerializedName("overview") val overview: String,
     @SerializedName("poster_path") val posterPath: String,
     @SerializedName("backdrop_path") val backdropPath: String,
-    @SerializedName("vote_average") val rating: Float,
+    @SerializedName("vote_average") val rating: Double,
     @SerializedName("release_date") val releaseDate: String
 )
